project_3/requirements.txt
glfw==2.7.0
jupyter==1.1.1
numpy==1.26.4
pyopengl==3.1.7
tabulate==0.9.0
